chore: update basic usage example
