+++ examples/next-basic/.eslintrc.json
