+++ examples/next-basic/next-env.d.ts
