+++ examples/next-basic/package.json
