+++ examples/next-basic/tsconfig.json
